requirements.txt
geopandas==1.0.1
numpy==2.2.4
rasterio==1.4.3
shapely==2.0.7
pyproj==3.7.1
docplex==2.29.241
cdsetool==0.2.13
python-dotenv==1.1.0
certifi>=2022.12.7
pandas==2.2.3
pillow==11.2.0
tqdm==4.67.1
requests==2.32.3
